chore(suite-native): remove module-devices
