+++ suite-native/assets/tsconfig.json
@@ -23,7 +23,6 @@
         { "path": "../ethereum-tokens" },
         { "path": "../fiat-rates" },
         { "path": "../formatters" },
-        { "path": "../module-devices" },
         { "path": "../module-settings" },
         { "path": "../navigation" },
         { "path": "../../packages/styles" }
