--- suite-native/module-devices/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "@suite-native/module-devices",
-    "version": "1.0.0",
-    "private": true,
-    "license": "See LICENSE.md in repo root",
-    "sideEffects": false,
-    "type": "module",
-    "main": "src/index",
-    "scripts": {
-        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "test:unit": "jest --passWithNoTests",
-        "type-check": "tsc --build"
-    },
-    "dependencies": {
-        "@reduxjs/toolkit": "1.9.5",
-        "@suite-common/suite-types": "workspace:*",
-        "@trezor/connect": "workspace:*",
-        "react": "18.2.0"
-    },
-    "devDependencies": {
-        "jest": "^26.6.3",
-        "typescript": "4.9.5"
-    }
-}
