--- suite-native/module-devices/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        {
-            "path": "../../suite-common/suite-types"
-        },
-        { "path": "../../packages/connect" }
-    ]
-}
