+++ suite-native/state/src/extraDependencies.ts
@@ -1,7 +1,7 @@
 import { ExtraDependencies } from '@suite-common/redux-utils';
 import { extraDependenciesMock } from '@suite-common/test-utils';
 import { enabledNetworks } from '@suite-native/config';
-import { selectDevices } from '@suite-native/module-devices';
+import { selectDevices } from '@suite-common/wallet-core';
 import { selectFiatCurrencyCode, setFiatCurrency } from '@suite-native/module-settings';
 import { PROTO } from '@trezor/connect';
 import { mergeDeepObject } from '@trezor/utils';
